Linux Installation Docs for Debian and CentOS
This patch updates the Linux installation docs to document both the
Debian as well as the RPM repository. This provides support for a wider
variety of distributions.

For now, the RPM repository is still marked as beta.

One thing this patch does that might be controversial (or maybe not?) is
the removal of the manual package installation. Apart maybe from
testing, there shouldn't be any good reason to choose that and I thought
it could also easily confuse people who then have trouble updating later
on. Maybe it's better to describe what we actually want people to do ;-)

That said, I don't have very strong feelings about this and I can
restore the section if anyone wants me to.