Cleaned the community section and added my ABStoolbox as an extra source (looking for opinions on whether it should be removed again).
Because many people do not know the audiobookshelf-web repository, and most things there are seen as official or finished projects, I started to curate a list of ABS projects and clients, even if the authors do not manually add them to the web. This should give the projects more exposure so they can be found, and it allows others to compare clients or projects on one page.